Rename data_base.rs to database.rs
